Issue #1719 - Improve handling of HTTP/2 queued requests.
Initial work that corrects the implementation of
HttpChannelOverHTTP2.onStreamTimeout().

Idle timeout have a special meaning in that they become a no-operation
if the application is dispatched but idle (neither reading nor writing).

HttpChannelOverHTTP2 now forwards the idle timeout to HttpInput, which
will only change its state if it is interested in reading.

HttpInput.consumeAll() has been modified to consume all input even if
it's already failed.